Don't look for patches to clean on a line item add

+++ fossbarcode/barcode/views.py
@@ -205,13 +205,14 @@ def detail(request, record_id):
                     
                 else:
                     # no patches specified, remove any that might be present
-                    patchdata = Patch_Files.objects.filter(frecord = foss_id)
-                    for patch in patchdata:
-                        try:
-                            pr.delete_file("patches/" + patch.path)
-                        except:
-                            error_message += "Failed to delete: " + patch.path + "<br>"
-                    patchdata.delete()
+                    if (mode != "Add Item"):
+                        patchdata = Patch_Files.objects.filter(frecord = foss_id)
+                        for patch in patchdata:
+                            try:
+                                pr.delete_file("patches/" + patch.path)
+                            except:
+                                error_message += "Failed to delete: " + patch.path + "<br>"
+                        patchdata.delete()
 
                 # update the master record "last updated"         
                 Product_Record.objects.filter(id = record_id).update(record_date = str(datetime.datetime.now()))
